requirements.txt
arrow==1.3.0
babel==2.16.0
binaryornot==0.4.4
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cookiecutter==2.6.0
cookiecutter-data-science==2.0.0
csscompressor==0.9.5
cycler==0.12.1
deeplay==0.1.3
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
ghp-import==2.1.0
htmlmin2==0.1.13
idna==3.10
Jinja2==3.1.5
jsmin==3.0.1
kiwisolver==1.4.8
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
material==0.1
matplotlib==3.10.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.49
mkdocs-material-extensions==1.3.1
mkdocs-minify-plugin==0.8.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
paginate==0.5.7
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
Pygments==2.18.0
pymdown-extensions==10.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scipy==1.15.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sympy==1.13.1
text-unidecode==1.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
torchmetrics==1.7.3
tqdm==4.67.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
watchdog==6.0.0
win32_setctime==1.2.0
